installer: upgrade to using InnoSetup v6.3.2
This version fixes a support function that we do not use
(`TStream.Seek`, see https://jrsoftware.org/files/is6-whatsnew.htm for
more details), but it's better to stay up to date than linger.

Incidentally, this also adds the Korean translations (which we also
don't use) which were apparently added in v6.3.0 but the file was not
committed at that time, apparently.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>