mingit: exclude new sshd executable
The `sshd.exe` file is already excluded from MinGit for quite a while.
As of OpenSSH 9.8, there is a new `sshd-session.exe`, which also needs
to be excluded.

This new file is part of an ongoing effort to split `sshd` further, see
https://github.com/openssh/openssh-portable/commit/03e3de416ed7c34faeb692967737be4a7bbe2eb5
for more details.

As a consequence, let's not just exclude `sshd-session.exe` piecemeal
fashion, but all `sshd-*.exe` files, whatever they be in the future.

This fixes the `check-for-missing-dlls` run that fails currently, see
https://github.com/git-for-windows/git-sdk-64/actions/runs/9754263596/job/26920893577#step:6:9
for an example.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>