ReleaseNotes: no longer mention Windows Explorer "bogus size" problem
It used to be the case that the size of the hard-linked "dashed
commands" was over-reported by Windows Explorer.

The warning about that has been obsolete for a while because newer
Windows Explorer versions do not miscount hard-links' sizes. In any
case, as of https://github.com/git-for-windows/git/pull/4252, Git for
Windows does not even install the dashed commands anymore. So let's drop
that warning.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>