Release Notes: clarify that Git for Windows is distributed with components that are not under the GPL
Git for Windows installs `git.exe`, which _is_ under the GNU Public
License v2. But it distributes many more components that are expected by
Git to be present (because Git expects to run on Linux, really) which
are not necessarily licensed under the same license. Since Git does not
link to those components (apart from a few, such as libcurl, which are
compatibly licensed), this is still respecting those licenses.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>