pacman-helper: make the script POSIX-compliant
In POSIX-compliant Unix shells, `$((` starts an arithmetic expression
that is ended by `))`.

Bash (being... Bash) allows `$((` not only to be an arithmetic
expression; It can alternatively be a short form of `$( (`, i.e. the
start of a shell expression that is to be executed and then interpolated
by its `stdout` where the shell expression immediately starts by opening
a subshell. Example:

  output="$( (cd / && ls -la) | tee >log.txt )"

This is valid by the POSIX standard; Removing the space between the two
open-parentheses would make it non-standard, where Bash would still run
it as intended but Dash, for example, wouldn't.

Let's allow executing this script using Dash.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>