installer: disable the "experimental" FSMonitor setting
The FSMonitor feature is no longer experimental, and hasn't been so for
quite a while. It also does not make sense to enable it on all
repositories, as the performance of using FSMonitor with small
repositories is potentially worse than without using the FSMonitor.

So let's disable the option, telling the user why.

This addresses https://github.com/git-for-windows/git/issues/4570, at
long last.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>