installer: hide the "experimental" FSMonitor setting by default
Seeing a disabled option that the user never enabled may be quite
confusing. So maybe just disabling it may be desirable only in scenarios
where users had enabled it?

So let's in fact not even show the FSMonitor option unless the user is
upgrading and had the option checked in the previous run.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>