installer: fix version comparison bug
This is a left-over from long ago, introduced in 490887923b (installer:
be prepared for Git tags when comparing versions, 2019-10-30).
Reconstructing my memory from that commit, it looks as if my intention
was to replace the faulty logic, but I left in that now-useless
conditional assignment of `Result` which is immediately overridden.

To help future me, let's just remove the unneeded code.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>