installer: skip Experimental Options page, if empty
As of a75602aa0 (installer: hide the "experimental" FSMonitor setting by
default, 2024-07-17), the FSMonitor option is hidden from the options
page unless the user had enabled it previously.

However, this then left the page blank (or: all options hidden), not a
good user experience. So let's skip it if it would appear blank.

This fixes https://github.com/git-for-windows/git/issues/5231.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>