installer: correctly label the Install button even when there is no experimental option
If all of the experimental options are hidden, we skip that page.
However, this is the last page of the wizard (at least if there are
no active processes that would block the installation), and that is
therefore the page where we re-label the "Next" button so that it says
"Install".

Therefore, if the experimental options page is hidden, we need to move
that button re-labeling to the preceding page.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>