please.sh build-mingw-w64-git: assume that all dependencies are installed
The idea of having `git-sdk-*` repositories that track full MSYS2
environments is not only to have a convenient and fast way to bootstrap
a build environment in which to build and release Git for Windows, it
also enables straight-forward processes for embargoed builds, where we
have to "time-travel" back to the state from which the preceding version
was built and assembled.

To guarantee that this time traveling continues to work, the build
process of a regular Git for Windows release must be prevented from
"conveniently" installing missing dependencies; If they are not part of
the `git-sdk-*` repository (or not in the `build-installers` sparse
checkout), that needs to be changed before releasing, so that the
embargoed release process can continue to work as intended.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>