Release Notes: announce that `git svn` will be phased out
It poses a substantial maintenance burden, what with the need not only
to keep maintaining Perl, but also subversion (and apr, apr-util,
and swig, and the Perl bindings of subversion).

At the same time, less and less `git svn` tickets (and no PRs worth
talking about) have trickled in.

Therefore it is a good time to say good-bye to `git svn` and send it off
to a cozy, relaxing retirement on the beach. It has served us well and
can now go on to enjoy life instead of work.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>