check-for-missing-dlls: stop using `BITNESS`
The `BITNESS` idea hails from the time when it was assumed that there
would only ever be a 32-bit and a 64-bit Windows, implying Intel/AMD
CPUs.

That has changed with the advent of ARM64, and we're slowly migrating
away from using `BITNESS`; This here commit is another step on that
journey.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>