make-file-list: stop being confused by `>=<ver>` in the `pactree` output
As of yesterday's sync of the i686 packages in MSYS2,
`mingw-w64-p11-kit` is listed with a `>=0.23.1` requirement. This
confused the `make-file-list.sh` script when it specifically wanted to
exclude `mingw-w64-p11-kit`'s files from being bundled.

Unbeknownst to me, this `>=` mishandling hid _another_ bug:
`mingw-w64-p11-kit>=0.23.1` is listed in the `git-sdk-64` repository as
a dependency of `mingw-w64-gnutls`, and that allowed it to slip through
the cracks and actually get included in 64-bit versions of Git for
Windows!

This finally also explains, at long last, the conundrum that puzzled me
so much in b7729d46 (make-file-list: do include p11-kit's dependencies,
2022-07-08) where the libffi/libtasn1 dependencies were needed in
git-sdk-64 but not in git-sdk-32: In 32-bit Git for Windows,
`/mingw32/bin/p11-kit.exe` was excluded but in 64-bit Git for Windows
`/mingw64/bin/p11-kit.exe` was included!

Let's make things consistent again and just ship the MINGW variant of
p11-kit in regular Git for Windows (but still omit it in MinGit).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>