make-file-list: work around 32-bit issues more surgically
As of 6f92a0eb (make-file-list: work around 32-bit issues, 2020-10-15),
we worked around a discrepancy between i686 and x86_64 variants of the
`apr` and the `file` package, excluding dependencies specifically for
i686 variants.

However, yesterday's big sync removed those discrepancies.

Let's use a `grep` invocation that is as straight-forward as it is hacky
as an indicator whether or not to special-case these dependencies in the
i686 case.

All of this is really only needed for this one last time when Git for
Windows v2.48.1 will ship with an i686 installer and portable Git;
Future versions of Git for Windows will only ship MinGit for that CPU
architecture, which does not have that problem.

Note that we also have to adjust the `please.sh create-sdk-artifact
build-installers` command for that hack to work there.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>