make-file-list: work around a libunistring version problem of gnupg
As of yesterday's big MSYS2 i686 sync, there is no longer an
`msys-unistring-2.dll` but an `msys-unistring-5.dll`.

However, gnupg (e.g. `dirmngr.exe`, via `msys-gnutls-30.dll`) depends on
the former file to be present.

Since the full i686 variant of Git for Windows is on its way out (we'll
be dropping the installer/portable Git builds for i686 after one last
full i686 release with Git for Windows v2.48.1, and MinGit does not
include gnupg), it isn't worth fretting too much about this; Simply
copying the latter to the former seems to work just fine.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>