git-extra: make `pkgver` slightly smarter
An update to the `pkgver` itself is not considered to be a vital change
to `git-extra`, therefore tip commits touching only that line are
excluded from the calculation of the package version.

However, running an `/updpkgsums` may very well update also a checksum
(which was exactly the case in
https://github.com/git-for-windows/build-extra/pull/591#issuecomment-2653057788)
and such a change should _likewise_ avoid advancing the `pkgver`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>