pacman-helper: drop the code to initialize the aarch64 repository
Back in 8e9b1571e (pacman-helper: be prepared to initialize the aarch64
Pacman repository, 2022-12-22), I had added code specifically to
initialize the aarch64 part of Git for Windows' Pacman repository, which
had not existed up to that point.

Now that the aarch64 repository is a thing, we can drop that code. In
fact, I could have dropped it a lot sooner, but hey, better late than
never.

It is actually a bit more pressing now because I want to transition
Git for Windows away from the Azure Blobs-backed Pacman repository, and
instead use https://github.com/git-for-windows/pacman-repo instead, via
a `/etc/pacman.conf` modification looking something like this:

  [git-for-windows-x86_64]
  Server = https://raw.githubusercontent.com/git-for-windows/pacman-repo/refs/heads/x86_64

  [git-for-windows-mingw32]
  Server = https://raw.githubusercontent.com/git-for-windows/pacman-repo/refs/heads/i686

For the transition period, I want to keep the Azure Blobs-backed
Pacman repository in sync but no longer make it the source of truth.
The code to initialize the aarch64 part of it would have made the
necessary changes unnecessarily complex. So away it goes.

This commit is best viewed with `-w`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>