pacman-helper quick_add: start transitioning deployments to GitHub
The idea is that the Pacman repository hosted on Azure Blobs keeps
functioning as before, for now, but that the GitHub repository at
https://github.com/git-for-windows/pacman-repo is now serviced as well.

It will serve as Git for Windows' official Pacman repository in the
future. The next step there is to do a couple of deployments, then
make the following adjustment to `/etc/pacman.conf` locally, and once
enough confidence has been gained, the post-install script of the
`mingw-w64-git-extra` package will learn to make this adjustment:

  diff --git a/etc/pacman.conf b/etc/pacman.conf
  index 99179ac6024..9555ae89af4 100644
  --- a/etc/pacman.conf
  +++ b/etc/pacman.conf
  @@ -70,17 +70,17 @@ LocalFileSigLevel = Optional
   # Server = https://repo.msys2.org/staging/
   # SigLevel = Never

  -[git-for-windows-aarch64]
  -Server = https://wingit.blob.core.windows.net/aarch64
  +[git-for-windows-clangarm64]
  +Server = https://raw.githubusercontent.com/git-for-windows/pacman-repo/refs/heads/aarch64

   [clangarm64]
   Include = /etc/pacman.d/mirrorlist.mingw

  -[git-for-windows]
  -Server = https://wingit.blob.core.windows.net/x86-64
  +[git-for-windows-x86_64]
  +Server = https://raw.githubusercontent.com/git-for-windows/pacman-repo/refs/heads/x86_64

   [git-for-windows-mingw32]
  -Server = https://wingit.blob.core.windows.net/i686
  +Server = https://raw.githubusercontent.com/git-for-windows/pacman-repo/refs/heads/i686

   [mingw32]
   Include = /etc/pacman.d/mirrorlist.mingw

What this means practically is that this commit introduces code that
will update the `x86_64`, `aarch64`, and `i686` branches in
`pacman-repo` to reflect the new deployment: new package version
archives will be added, the package database updated, and any overridden
package versions' archives will be removed.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>