pacman-helper quick_add: use new-style database filenames locally
The plan is to mirror each Pacman repository deployment to a new GitHub
Release in the `git-for-windows/pacman-repo` repository. To allow for
that, the database filenames must be unique, which means that we will
add the `-<architecture>` infix to the `git-for-windows.db` files (and
likewise the `.files` ones).

This requires the MINGW-only `git-for-windows-aarch64` to be renamed to
`git-for-windows-clangarm64` (which it should have been named in the
first place, in line with the other two MINGW-only databases whose
filenames reflect the down-cased `MSYSTEM`).

In preparation for modifying the code to upload these GitHub Releases,
adapt the existing code to use the new-style naming at least locally.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>