wingit-snapshot-helper: drop the `add-snapshot` sub-command
I recently transitioned Git for Windows' snapshot builds away from being
hosted on Azure Blobs. Instead, they are now hosted at
https://github.com/git-for-windows/git-snapshots/releases, with the
overview page as a GitHub Page at
https://gitforwindows.org/git-snapshots

Which means that the snapshot builds are not even uploaded to Azure
Blobs anymore, and the corresponding code to upload them can be dropped,
which I hereby do.

Granted, the script's filename is now a bit misleading because the only
operations it supports now are related to Git for Windows' Pacman
repository that is hosted on Azure Blobs.

But I am currently embarking on the project to transition also the
Pacman repository to GitHub, concretely to the branches of
https://github.com/git-for-windows/pacman-repo whose branch names
reflect the corresponding CPU architecture. Once that transition is
done, the entire `wingit-snapshot-helper.sh` script can be dropped, so
I'll not bother with renaming it for that (hopefully short) time window.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>