pacman-helper: drop most sub-commands
In the early days of Git for Windows v2.x, when automation meant running
scripts in a clean VM I kept on my laptop, the Pacman repository was
simply a remote mirror of a local directory tree.

Over the years, many things changed. We transitioned from Bintray to
Azure Blobs (because the former was woefully unreliable and also simply
_stopped_ serving _any_ file once we hit a relatively low quota), we
transitioned building Pacman packages first to Azure Pipelines and then
to GitHub Actions.

This also meant that the Pacman repository no longer
had a local source of truth but instead a remote one, and the automation
required the ability to add a package without initializing a local
mirror on the Action runners: The `quick_add` sub-command was born, see
b408f99559 (pacman-helper: add a quick mode, 2019-11-27).

This `quick_add` operation merely downloads the package databases, runs
`repo_add` with the packages to add, and then uploads the package
databases along with the new package version's archives.

Ever since that operation was introduced, I haven't had the need for any
of the other ways to upload or verify packages, and it is time to retire
the code associated with those no-longer-used functionality.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>