pacman-helper quick_add: verify that the old/new Pacman repos are in sync
Git for Windows' current Pacman repository is hosted on Azure Blobs.

The intended new Pacman repository is actually a Git repository hosted
on GitHub: https://github.com/git-for-windows/pacman-repo

The `pacman-helper.sh` script was taught in the preceding commit to
update both synchronously, not just the Azure Blobs one.

To ensure that this is all working as intended, let's make extra certain
that these two Pacman repositories have not diverged before making any
changes.

This uncovered a bug where the `sanitize_db` function would update the
signature file even if that was not desired nor necessary. We fix that
here because said "are we in sync?" check would otherwise fail.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>