pacman-helper quick_add: allow for concurrent deployments
It is completely normal to have multiple concurrent deployments, say,
`mingw-w64-curl` and `mingw-w64-git-extra`. This means that
`pacman-helper quick_add` might update a branch locally that cannot be
pushed because another deployment has already pushed the branch in the
meantime.

This is not a problem, as long as the deployments touch separate
packages. In that instance, we simply need to revert the changes to the
package database, then rebase to the latest commit from the public
`pacman-repo` repository, and try again.

Note: Currently this situation cannot even occur because we're doing the
dual deployment both to Azure Blobs as well as to `pacman-repo`, and the
former is protected against concurrent deployments via the lease
mechanism.

Nevertheless, it is a good idea to make the code ready for the day when
we no longer deploy to Azure Blobs.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>