make-file-list.sh: future-proof the `pactree` output parsing
In c9f04bbd9 (make-file-list: stop being confused by `>=<ver>` in the
`pactree` output, 2025-02-10), this script was adjusted to account for
`pactree` sometimes specifying version ranges of some dependencies.

However, this change was incomplete! It only handles the version ranges
that are _currently_ found in Git for Windows, but other version ranges
are equally valid.

What I came up with to parse a dependency in msys2/pacdb
(https://github.com/msys2/pacdb) is:

```source-python
_DEPENDRE = re.compile(r'([^<>=]+)(?:(<=|>=|<|>|=)(.*))?')
```

That might be more complicated than we want to use here though. (That
gets you 3 match groups: the package name, the relational operator, and
the version).

We might just as well future-proof it by a simpler solution that
nevertheless catches more package version checks.

Signed-off-by: Jeremy Drake <github@jdrake.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>