pacman-helper quick_add: fix gpg verification in `sanitize_db`
In 94dde8c2b (pacman-helper quick_add: verify that the old/new Pacman
repos are in sync, 2025-02-16), I added a call to `call_gpg --verify`.
But that fails in the regular Git for Windows workflows because there,
`gpg.program=gnupg-with-gpgkey.sh` is configured and hence the
`call_gpg` function calls that script. And that script really only helps
with signing, not with verifying (but fails instead).

So let's call `gpg --verify` directly instead.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>