pacman-helper quick_add: try harder to stay in sync with pacman-repo
The worktrees are initialized before the lease is acquired, therefore if
the lease could not be acquired right away, chances are that the
worktree will be stale once the lease _can_ be acquired.

This was the case in
https://github.com/git-for-windows/git-for-windows-automation/actions/runs/13378287807/job/37362013059#step:20:71
where a concurrent deployment succeeded and updated both Pacman
repositories in the meantime, and the worktree was hence outdated and
the up to date check failed thusly:

  git-for-windows-i686.db.tar.xz: needs update
  git-for-windows-i686.db.tar.xz.sig: needs update
  git-for-windows-i686.files.tar.xz: needs update
  git-for-windows-i686.files.tar.xz.sig: needs update
  The package databases in i686 differ between Azure Blobs and pacman-repo

Let's just update the branch just before downloading the package
databases from Azure Blobs, to give it a chance to be in sync.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>