pacman-helper quick_remove: allow specifying package names
Instead of filenames, it is much more natural to specify the package
name. One caveat is that the package name must also contain the CPU
architecture in that case, but it can leave out the version number.

For example, instead of

  pacman-helper.sh quick_remove cyrus-sasl-2.1.28-2-i686.pkg.tar.xz

one can now write

  pacman-helper.sh quick_remove cyrus-sasl-i686

To make this safer with regards to package names that end in
`-<digit>[...]` (like `msys2-runtime-3.3`), either the full version has
to be provided, including pkgrel, or the version has to be omitted.

Co-authored-by: Jeremy Drake <github@jdrake.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>