pacman-helper.sh quick_remove: skip uploading package files
The (dual) strategy of the new Pacman repository is to not only publish
the latest state in the `x86_64`, `aarch64` and `i686` branches, but
also to upload the files in the latest update to a GitHub Release.

Naturally, when adding a new package version, the package archive and
its PGP signature need to be included in that Release.

However, when _removing_ a package, the corresponding package archive
should of course be omitted ;-)

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>