pacman-helper.sh: set release name to be the same as tagname
I noticed the latest release created picked up its name from the latest commit to the main branch in the pacman-repo repository, namely "2025-02-26T22-44-26.725080300Z: Merge pull request #1 from git-for-windows/populate-the-readme".  Given that it's very unlikely that the commit(s) on the main branch have anything to do with the releases, just specify the name to be the same as the tagname.

Signed-off-by: jeremyd2019 <github@jdrake.com>