pacman-helper.sh: compose release body/notes
Do some trivial sed escaping of characters that might cause issues in
JSON.  This is not comprehensive, but the text in question is controlled
by this script and package names and versions being published, so we can
reasonably be assured of a limited set of characters being used.  (In
fact, it's probably unnecessary to escape quotes and backslashes).   

Co-authored-by: Johannes Schindelin <johannes.schindelin@gmx.de>
Signed-off-by: Jeremy Drake <github@jdrake.com>