ReleaseNotes: stop mentioning certificates that aren't bundled
Ever since the switch from MSys/MinGW to MSYS2/MINGW in 2015, i.e.
almost 10 years ago, Git for Windows hasn't bundled the certificates
that were specifically mentioned in the release notes.

They were introduced via https://github.com/msysgit/msysgit/issues/37,
which is _such_ a blast from a more pleasant past. But since the
certificates aren't included any more, we should not claim that we do in
the release notes, either.

It's somewhat funny that nobody realized this in almost a decade.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>