installer: when uninstalling, do remove `cmd`
In dcd28b3272 (installer: add the git-lfs wrapper to /cmd, 2018-02-21),
we added code to install the `git-lfs` wrapper to the `cmd` directory,
and we even added code to remove it when uninstalling.

But we forgot to actually remove the `cmd` directory itself (that then
should be empty). This has been pointed out in
https://github.com/git-for-windows/git/issues/5492#issuecomment-2732669971

Let's fix that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>