installer: when uninstalling, really remove the `etc` directory
In 4b546cc56b (installer: allow Pseudo Console support to be
re-disabled, 2020-07-02), we added support for the
`/etc/git-bash.config` file, including a line to uninstall that file
together with the rest of Git for Windows.

But we forgot to move the line to remove the `/etc` directory as needed.
As a consequence, uninstalling Git for Windows leaves behind the `etc`
directory by mistake.

While at it, remove the line in the 32-bit-specific section that wanted
to remove the `etc` directory. It should have been moved outside the
32-bit-specific section, anywa, but 8911bd5832 (installer: record
install options in `/etc/install-options.txt`, 2015-10-01) duplicated a
line later on after ensuring that `install-options.txt` would be
removed.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>