installer: do remove the system config when uninstalling
When uninstalling Git for Windows, `gitconfig` remains, as pointed out in
https://github.com/git-for-windows/git/issues/5492#issuecomment-2733104264

The reason for this is a regression I introduced in c0fb1e5275
(installer: re-generate the system config instead of copying it,
2019-02-19): before that change, `/etc/gitconfig` was copied regularly,
and InnoSetup would of course know to remove it as part of uninstalling
the software. After that change, however, the system config is
(re-)generated dynamically, and InnoSetup does not know that it needs to
be removed when uninstalling.

Let's tell InnoSetup that it should delete that file when uninstalling.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>