installer: ensure that `/dev/` exists before populating it
In https://github.com/git-for-windows/git/issues/5492 a problem was
reported where the installer could not access `/dev/` because it did not
exist yet.

It is unclear why that directory did not exist yet (it is explicitly
mentioned in the `[Dirs]` section and should therefore have been created
before the `ssPostInstall` step), and it would appear that usually it
exists at this point.

Let's just ensure that the installer creates `/dev/` before populating
it with special symbolic links expected by the Cygwin/MSYS2 runtime.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>