ever-green: avoid non-POSIX shell pattern
The following construct relies on a non-compliant shell pattern:

  case "$number" in
  *[^0-9]*) echo "Not a number";;
  esac

The reason is that the `[^...]` construct is not actually correct. Per
https://pubs.opengroup.org/onlinepubs/9799919799/utilities/V3_chap02.html#tag_19_14_01:

  A <left-square-bracket> shall introduce a bracket expression if the
  characters following it meet the requirements for bracket expressions
  stated in XBD 9.3.5 RE Bracket Expression, except that the
  <exclamation-mark> character ('!') shall replace the <circumflex>
  character ('^') in its role in a non-matching list in the regular
  expression notation. A bracket expression starting with an unquoted
  <circumflex> character produces unspecified results.

Therefore the correct way to write this character class is: `[!0-9]`.

Let's use that here because the `ubuntu-24.04` GitHub Actions/Azure
Pipeline runners now ship with dash v0.5.12 (which is used as /bin/sh on
Ubuntu), and unlike version 0.5.11, that newer version is much more
stringent when it comes to above-mentioned `[^...]` construct (read:
this script no longer worked correctly in Git for Windows' automation as
run in an Azure Pipeline).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>