git-extra: switch to the new Pacman repo
For a couple of months now, to precise: since
https://github.com/git-for-windows/build-extra/pull/595 (Let
`pacman-helper` deploy also to the GitHub repo
`git-for-windows/pacman-repo`), every deployment to Git for Windows'
Pacman repository is mirrored to a new location:
https://github.com/git-for-windows/pacman-repo/

The idea set forth in one of the commits in that PR, 3fca5c703
(pacman-helper quick_add: start transitioning deployments to GitHub,
2025-02-16), is to switch to the new location eventually, avoiding the
dependency on my personal Azure account (where the old location
resides).

I had meant to transition a lot quicker, but procrastinated on this for
such a long time that I forgot. But today is the day!

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>