install.iss: add Microsoft Edit as an editor option
Add the newly re-written and open-source 'Edit' terminal-based text
editor, inspired by the old MS-DOS Editor.

While the edit.exe program is also available as a simple zip download, or
via the WinGet package manager for Windows, it is also distributed via
Windows Update on Windows 11 24H2 and newer. Since that's the only system-wide
install with a stable path, we currently only support that install method.

Signed-off-by: Matthew John Cheetham <mjcheetham@outlook.com>
Signed-off-by: Matthias Aßhauer <mha1993@live.de>