install.iss: add Microsoft Edit as an editor option
Add the newly re-written and open-source 'Edit' terminal-based text
editor, inspired by the old MS-DOS Editor.

Note that at this time the edit.exe program is only available as a
simple zip download, or via the WinGet package manager for Windows.
Since we can only rely on a stable path for the WinGet package manager
we should look there (per-user).

Signed-off-by: Matthew John Cheetham <mjcheetham@outlook.com>