installer: test any custom editor on a _writable_ file
When choosing a custom editor, there is the convenient "Test Custom
Editor" button. When clicking it, a file is written and the specified
editor is invoked to edit that file. The file is then deleted.

The contents of that file specifically encourage the user to make
changes and write the file.

Unfortunately, due to security concerns InnoSetup nowadays restricts the
permissions of its temporary directory, which means that the file (which
is written by the elevated installer process) is not writable by the
(non-elevated) custom editor.

Let's just write the file _next_ to that temporary directory, which is
the user's TEMP directory and therefore writable.

This fixes https://github.com/git-for-windows/git/issues/5618.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>