git-extra: rename the local copy of the Pacman database as needed
Recently we switched from the old Pacman repository that was hosted on
Azure Blobs in my personal Azure account to the new one that is hosted
at https://github.com/git-for-windows/pacman-repo.

To make that work, we needed to rename the Pacman database from the
generic `git-for-windows.db` to `git-for-windows-<arch>.db` where
`<arch>` is the architecture of the current MSYS2 runtime.

Unfortunately, the `sync` workflow of the `git-sdk-*` repositories was
not quite prepared for that rename and could no longer find the
`mingw-w64-git-extra` package to install.

Side note: This did not affect the `git-sdk-32` repository because that
is stuck with an older Pacman version that did not quite care about the
current version of the `mingw-w64-git-extra` package when being asked
to reinstall it.

Let's make sure that the database file is renamed as appropriate.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>